+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 Theme-Solutions theme framework changes are logged here using <a href="http://semver.org/">Semantic Versioning</a>.
 
 
+## 1.2.1 (2019-07-07) ##
+* Changed settings functions to use the WordPress-Soutions plugin.
+* Improved automatic updates configuration.
+* Updated all version numbers to be the same.
+
 ## 1.1.1 (2019-04-11) ##
 * Added remove 'Gutenberg CSS' option.
 * Updated all version numbers to be the same.